Tests done for the User!
Now the tests for the `User` model and its `UserPropertyValidator`. There were a minor change done to the way a user is validated, so now instead of returning at the first point of failure, it continues to check the rest as well. Then just returning a list of errors, so the user is able to fix all the potential errors in one swoop.